move network interceptor examples to bidi file and update docs references
